requirements.txt
argparse
datasets
fasttext
numpy
pandas
pyarrow
sklearn
torch
torchvision
torchaudio
transformers[torch]
sentence_transformers
typing-inspect
typing_extensions
spacy
pyyaml

https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0.tar.gz#egg=fr_core_news_sm
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz#egg=en_core_web_sm